[chore/#230] myProfileMapper 분리
